requirements.txt
accelerate==0.34.2
anyio==4.6.0
appdirs==1.4.4
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boilerpy3==1.0.7
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.1
docopt==0.6.2
Events==0.5
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.7.2
farm-haystack==1.26.3
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
importlib-metadata==4.6.4
inflect==7.4.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==23.5.0
langdetect==1.0.9
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-imports==0.3.1
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
netifaces==0.11.0
networkx==3.3
nltk==3.9.1
num2words==0.5.13
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parsivar==0.2.3.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
posthog==3.6.6
prompt_toolkit==3.0.47
prompthub-py==4.0.0
protobuf==5.28.2
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==1.10.18
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
quantulum3==0.9.2
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-cache==0.9.8
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
SecretStorage==3.3.1
sentence-transformers==3.0.0
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-Utils==0.41.2
sseclient-py==1.8.0
stack-data==0.6.3
sympy==1.13.3
systemd-python==234
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
triton==3.0.0
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
url-normalize==1.4.3
urllib3==2.2.3
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
yarg==0.1.9
zipp==1.0.0
